Rename exec to structex
