+++ src/editor/actions.rs
@@ -5,12 +5,12 @@ use crate::{
     config_handle,
     dot::{Cur, Dot, Range, TextObject},
     editor::{Editor, MbSelector, MiniBufferSelection},
-    exec::{Addr, Address, Program},
     fsys::LogEvent,
     key::{Arrow, Input},
     lsp::Coords,
     mode::Mode,
     plumb::{MatchOutcome, PlumbingMessage},
+    structex::{Addr, Address, Program},
     system::System,
     ui::{StateChange, UserInterface},
     util::gen_help_docs,
