+++ src/editor/mod.rs
@@ -5,13 +5,13 @@ use crate::{
     config::Config,
     config_handle, die,
     dot::TextObject,
-    exec::{Addr, Address},
     fsys::{AdFs, LogEvent, Message, Req},
     input::Event,
     key::{Arrow, Input},
     lsp::{LspManager, LspManagerHandle},
     mode::{Mode, modes},
     plumb::PlumbingRules,
+    structex::{Addr, Address},
     system::{DefaultSystem, System},
     term::CurShape,
     ui::{Layout, SCRATCH_ID, StateChange, Ui, UserInterface},
